Move `*Selector` from `re_chunk_store` to `re_sorbet` (#9851)
### Related

* Part of #9837

### What

Move `{Column|ComponentColumn|TimelineColumn}Selector` to `re_sorbet`
where they belong (alongside the `*Descriptor` crowd).
