+++ crates/viewer/re_view_dataframe/Cargo.toml
@@ -31,6 +31,7 @@ re_format.workspace = true
 re_log_types.workspace = true
 re_log.workspace = true
 re_renderer.workspace = true
+re_sorbet.workspace = true
 re_tracing.workspace = true
 re_types_core.workspace = true
 re_types.workspace = true
